Rename optimize_move to optimize_rewrite
The module now handles select lowering, move insertion, and value copy
type collection — not just moves.
