+++ wado-compiler/src/optimize_rewrite.rs
